Changed default navigation in double quads
